Add a "reverse path" link when displaying the shortest paths from A to B.
It seems that the "reverse path" link is useless, because the picture and the tables
will display exactly the same thing for the "A → B" shortest paths and for the
"B → A" shortest paths. There is still a difference. In the webpage displaying
the "A → B" shortest paths, the links from the node table refer to the
webpages displaying the "A → X" shortest paths. In the webpage displaying
the "B → A" shortest paths, the links from the node table refer to the
webpages displaying the "B → X" shortest paths. Not a big deal, but it helps
the user's navigation.